[Fixes #222] Align geonode-project 3.3.x structure to master branch
