+++ .env
@@ -7,18 +7,16 @@ DOCKER_API_VERSION="1.24"
 BACKUPS_VOLUME_DRIVER=local
 
 C_FORCE_ROOT=1
-IS_CELERY=false
 FORCE_REINIT=false
-
-SITEURL=http://localhost/
-ALLOWED_HOSTS=['django',]
+INVOKE_LOG_STDOUT=true
 
 # LANGUAGE_CODE=pt
 # LANGUAGES=(('en','English'),('pt','Portuguese'))
 
-GEONODE_INSTANCE_NAME=geonode
 DJANGO_SETTINGS_MODULE={{project_name}}.settings
-UWSGI_CMD=uwsgi --ini /usr/src/{{project_name}}/uwsgi.ini
+GEONODE_INSTANCE_NAME=geonode
+GEONODE_LB_HOST_IP=
+GEONODE_LB_PORT=
 
 # #################
 # backend
@@ -41,30 +39,18 @@ DEFAULT_BACKEND_DATASTORE=datastore
 BROKER_URL=amqp://guest:guest@rabbitmq:5672/
 ASYNC_SIGNALS=True
 
-# #################
-# geoserver
-# #################
-GEOSERVER_WEB_UI_LOCATION=http://localhost/geoserver/
-GEOSERVER_PUBLIC_LOCATION=http://localhost/geoserver/
-GEOSERVER_LOCATION=http://geoserver:8080/geoserver/
-GEOSERVER_ADMIN_USER=admin
-GEOSERVER_ADMIN_PASSWORD=geoserver
+SITEURL=http://localhost/
 
-OGC_REQUEST_TIMEOUT=30
-OGC_REQUEST_MAX_RETRIES=1
-OGC_REQUEST_BACKOFF_FACTOR=0.3
-OGC_REQUEST_POOL_MAXSIZE=10
-OGC_REQUEST_POOL_CONNECTIONS=10
-
-# Java Options & Memory
-ENABLE_JSONP=true
-outFormat=text/javascript
-GEOSERVER_JAVA_OPTS=-Djava.awt.headless=true -Xms2G -Xmx4G -XX:+UnlockDiagnosticVMOptions -XX:+LogVMOutput -XX:LogFile=/var/log/jvm.log -XX:PerfDataSamplingInterval=500 -XX:SoftRefLRUPolicyMSPerMB=36000 -XX:-UseGCOverheadLimit -XX:+UseConcMarkSweepGC -XX:+UseParNewGC -XX:ParallelGCThreads=4 -Dfile.encoding=UTF8 -Djavax.servlet.request.encoding=UTF-8 -Djavax.servlet.response.encoding=UTF-8 -Duser.timezone=GMT -Dorg.geotools.shapefile.datetime=false -DGEOSERVER_CSRF_DISABLED=true -DPRINT_BASE_URL=http://geoserver:8080/geoserver/pdf -DALLOW_ENV_PARAMETRIZATION=true -Xbootclasspath/a:/usr/local/tomcat/webapps/geoserver/WEB-INF/lib/marlin-0.9.3-Unsafe.jar -Dsun.java2d.renderer=org.marlin.pisces.MarlinRenderingEngine
+ALLOWED_HOSTS=['django', '*']
 
 # Data Uploader
 DEFAULT_BACKEND_UPLOADER=geonode.importer
 TIME_ENABLED=True
 MOSAIC_ENABLED=False
+HAYSTACK_SEARCH=False
+HAYSTACK_ENGINE_URL=http://elasticsearch:9200/
+HAYSTACK_ENGINE_INDEX_NAME=haystack
+HAYSTACK_SEARCH_RESULTS_PER_PAGE=200
 
 # #################
 # Jenkins
@@ -100,6 +86,26 @@ LETSENCRYPT_MODE=disabled
 RESOLVER=127.0.0.11
 
 # #################
+# geoserver
+# #################
+GEOSERVER_WEB_UI_LOCATION=http://localhost/geoserver/
+GEOSERVER_PUBLIC_LOCATION=http://localhost/geoserver/
+GEOSERVER_LOCATION=http://geoserver:8080/geoserver/
+GEOSERVER_ADMIN_USER=admin
+GEOSERVER_ADMIN_PASSWORD=geoserver
+
+OGC_REQUEST_TIMEOUT=30
+OGC_REQUEST_MAX_RETRIES=1
+OGC_REQUEST_BACKOFF_FACTOR=0.3
+OGC_REQUEST_POOL_MAXSIZE=10
+OGC_REQUEST_POOL_CONNECTIONS=10
+
+# Java Options & Memory
+ENABLE_JSONP=true
+outFormat=text/javascript
+GEOSERVER_JAVA_OPTS=-Djava.awt.headless=true -Xms2G -Xmx4G -XX:+UnlockDiagnosticVMOptions -XX:+LogVMOutput -XX:LogFile=/var/log/jvm.log -XX:PerfDataSamplingInterval=500 -XX:SoftRefLRUPolicyMSPerMB=36000 -XX:-UseGCOverheadLimit -XX:+UseConcMarkSweepGC -XX:+UseParNewGC -XX:ParallelGCThreads=4 -Dfile.encoding=UTF8 -Djavax.servlet.request.encoding=UTF-8 -Djavax.servlet.response.encoding=UTF-8 -Duser.timezone=GMT -Dorg.geotools.shapefile.datetime=false -DGEOSERVER_CSRF_DISABLED=true -DPRINT_BASE_URL=http://geoserver:8080/geoserver/pdf -DALLOW_ENV_PARAMETRIZATION=true -Xbootclasspath/a:/usr/local/tomcat/webapps/geoserver/WEB-INF/lib/marlin-0.9.3-Unsafe.jar -Dsun.java2d.renderer=org.marlin.pisces.MarlinRenderingEngine
+
+# #################
 # Security
 # #################
 # Admin Settings
@@ -127,7 +133,7 @@ DEFAULT_FROM_EMAIL='GeoNode <no-reply@geonode.org>'
 # Session/Access Control
 LOCKDOWN_GEONODE=False
 CORS_ORIGIN_ALLOW_ALL=True
-X_FRAME_OPTIONS=SAMEORIGIN
+X_FRAME_OPTIONS="SAMEORIGIN"
 SESSION_EXPIRED_CONTROL_ENABLED=True
 DEFAULT_ANONYMOUS_VIEW_PERMISSION=True
 DEFAULT_ANONYMOUS_DOWNLOAD_PERMISSION=True
@@ -224,3 +230,6 @@ LDAP_GROUP_PROFILE_MEMBER_ATTR=uniqueMember
 # CELERY__LOG_LEVEL="INFO"
 # CELERY__LOG_FILE="/var/log/celery.log"
 # CELERY__WORKER_NAME="worker1@%h"
+
+# PostgreSQL
+POSTGRESQL_MAX_CONNECTIONS=200
