+++ Dockerfile
@@ -28,45 +28,41 @@ RUN apt-get update && apt-get install -y \
     firefox-esr \
     --no-install-recommends && rm -rf /var/lib/apt/lists/*
 
-
-# add bower and grunt command
-COPY . /usr/src/{{project_name}}/
-WORKDIR /usr/src/{{project_name}}
-
-COPY monitoring-cron /etc/cron.d/monitoring-cron
-RUN chmod 0644 /etc/cron.d/monitoring-cron
-RUN crontab /etc/cron.d/monitoring-cron
-RUN touch /var/log/cron.log
-RUN service cron start
-
-COPY wait-for-databases.sh /usr/bin/wait-for-databases
-RUN chmod +x /usr/bin/wait-for-databases
-RUN chmod +x /usr/src/{{project_name}}/tasks.py \
-    && chmod +x /usr/src/{{project_name}}/entrypoint.sh
-
-COPY celery.sh /usr/bin/celery-commands
-RUN chmod +x /usr/bin/celery-commands
-
-COPY celery-cmd /usr/bin/celery-cmd
-RUN chmod +x /usr/bin/celery-cmd
-
 # Prepraing dependencies
 RUN apt-get update && apt-get install -y devscripts build-essential debhelper pkg-kde-tools sharutils
 # RUN git clone https://salsa.debian.org/debian-gis-team/proj.git /tmp/proj
 # RUN cd /tmp/proj && debuild -i -us -uc -b && dpkg -i ../*.deb
 
 # Install pip packages
-RUN pip install pip --upgrade
-RUN pip install --upgrade --no-cache-dir  --src /usr/src -r requirements.txt \
+RUN pip install pip --upgrade \
     && pip install pygdal==$(gdal-config --version).* \
-    && pip install flower==0.9.4
-
-RUN pip install --upgrade  -e .
+        flower==0.9.4
 
 # Activate "memcached"
 RUN apt install -y memcached
 RUN pip install pylibmc \
     && pip install sherlock
+
+# add bower and grunt command
+COPY src /usr/src/{{project_name}}/
+WORKDIR /usr/src/{{project_name}}
+
+COPY src/monitoring-cron /etc/cron.d/monitoring-cron
+RUN chmod 0644 /etc/cron.d/monitoring-cron
+RUN crontab /etc/cron.d/monitoring-cron
+RUN touch /var/log/cron.log
+RUN service cron start
+
+COPY src/wait-for-databases.sh /usr/bin/wait-for-databases
+RUN chmod +x /usr/bin/wait-for-databases
+RUN chmod +x /usr/src/{{project_name}}/tasks.py \
+    && chmod +x /usr/src/{{project_name}}/entrypoint.sh
+
+COPY src/celery.sh /usr/bin/celery-commands
+RUN chmod +x /usr/bin/celery-commands
+
+COPY src/celery-cmd /usr/bin/celery-cmd
+RUN chmod +x /usr/bin/celery-cmd
 
 # Install "geonode-contribs" apps
 RUN cd /usr/src; git clone https://github.com/GeoNode/geonode-contribs.git -b master
@@ -74,4 +70,11 @@ RUN cd /usr/src; git clone https://github.com/GeoNode/geonode-contribs.git -b ma
 RUN cd /usr/src/geonode-contribs/geonode-logstash; pip install --upgrade  -e . \
     cd /usr/src/geonode-contribs/ldap; pip install --upgrade  -e .
 
-ENTRYPOINT /usr/src/{{project_name}}/entrypoint.sh
+RUN pip install --upgrade --no-cache-dir  --src /usr/src -r requirements.txt
+RUN pip install --upgrade  -e .
+
+# Export ports
+EXPOSE 8000
+
+# We provide no command or entrypoint as this image can be used to serve the django project or run celery tasks
+# ENTRYPOINT /usr/src/{{project_name}}/entrypoint.sh
