--- docker-compose.override.yml
@@ -1,23 +1,0 @@
-version: '2.2'
-services:
-
-  django:
-    build: .
-    # Loading the app is defined here to allow for
-    # autoreload on changes it is mounted on top of the
-    # old copy that docker added when creating the image
-    volumes:
-      - '.:/usr/src/{{project_name}}'
-      # - '/data/backup_restore:/backup_restore'  # Link to local volume in the HOST
-
-  # celery:
-  #   volumes:
-  #     - '/data/backup_restore:/backup_restore'  # Link to local volume in the HOST
-
-  # geoserver:
-  #   volumes:
-  #     - '/data/backup_restore:/backup_restore'  # Link to local volume in the HOST
-
-  # jenkins:
-  #   volumes:
-  #     - '/data/backup_restore:/backup_restore'  # Link to local volume in the HOST
