+++ docker-compose.yml
@@ -1,148 +1,140 @@
-version: '2.2'
+version: '3.4'
 services:
-  db:
-    image: geonode/postgis:13
-    restart: unless-stopped
-    container_name: db4${COMPOSE_PROJECT_NAME}
-    stdin_open: true
-    # tty: true
-    labels:
-        org.geonode.component: db
-        org.geonode.instance.name: geonode
-    volumes:
-      - dbdata:/var/lib/postgresql/data
-      - dbbackups:/pg_backups
-    env_file:
-      - .env
 
-  rabbitmq:
-    image: rabbitmq
-    restart: unless-stopped
-    container_name: rabbitmq4${COMPOSE_PROJECT_NAME}
-    stdin_open: true
-    tty: true
-    labels:
-        org.geonode.component: rabbitmq
-        org.geonode.instance.name: geonode
-        org.geonode.deployment.name: ${COMPOSE_PROJECT_NAME}
+# Common Django template for GeoNode and Celery services below
+x-common-django:
+  &default-common-django
+  image: ${COMPOSE_PROJECT_NAME}_django:latest
+  restart: on-failure
+  env_file:
+    - .env
+  volumes:
+    # - '.:/usr/src/{{project_name}}'
+    - statics:/mnt/volumes/statics
+    - geoserver-data-dir:/geoserver_data/data
+    - backup-restore:/backup_restore
+    - data:/data
+    - tmp:/tmp
 
-  geoserver:
-    image: geonode/geoserver:2.18.2
-    build: ./docker/geoserver/
-    restart: unless-stopped
-    container_name: geoserver4${COMPOSE_PROJECT_NAME}
-    stdin_open: true
-    # tty: true
-    labels:
-        org.geonode.component: geoserver
-        org.geonode.instance.name: geonode
-    depends_on:
-      - db
-      - data-dir-conf
-    volumes:
-      - statics:/mnt/volumes/statics
-      - geoserver-data-dir:/geoserver_data/data
-      - backup-restore:/backup_restore
-      - data:/data
-      - tmp:/tmp
-    env_file:
-      - .env
+services:
 
+  # Our custom django application. It includes Geonode.
   django:
-    restart: unless-stopped
-    build: .
+    << : *default-common-django
+    build:
+      context: ./
+      dockerfile: Dockerfile
     container_name: django4${COMPOSE_PROJECT_NAME}
-    stdin_open: true
-    # tty: true
-    labels:
-      org.geonode.component: django
-      org.geonode.instance.name: geonode
-    depends_on:
-      - db
-      - rabbitmq
-      - data-dir-conf
-    volumes:
-      - statics:/mnt/volumes/statics
-      - geoserver-data-dir:/geoserver_data/data
-      - backup-restore:/backup_restore
-      - data:/data
-      - tmp:/tmp
-    env_file:
-      - .env
+    healthcheck:
+      test: "curl --fail --silent --write-out 'HTTP CODE : %{http_code}\n' --output /dev/null http://127.0.0.1:8001/"
+      interval: 60s
+      timeout: 10s
+      retries: 1
+      start_period: 60s
     environment:
-      UWSGI_CMD: uwsgi --ini /usr/src/{{project_name}}/uwsgi.ini
-      IS_CELERY: 'False'
+      - IS_CELERY=False
+    entrypoint: ["/usr/src/{{project_name}}/entrypoint.sh"]
+    command: "uwsgi --ini /usr/src/{{project_name}}/uwsgi.ini"
 
+  # Celery worker that executes celery tasks created by Django.
   celery:
-    restart: unless-stopped
-    #build: .
+    << : *default-common-django
     image: ${COMPOSE_PROJECT_NAME}_django:latest
     container_name: celery4${COMPOSE_PROJECT_NAME}
-    stdin_open: true
-    # tty: true
-    labels:
-      org.geonode.component: celery
-      org.geonode.instance.name: geonode
     depends_on:
       - django
-      - db
-      - rabbitmq
-      - data-dir-conf
-    volumes:
-      - statics:/mnt/volumes/statics
-      - geoserver-data-dir:/geoserver_data/data
-      - backup-restore:/backup_restore
-      - data:/data
-      - tmp:/tmp
-    env_file:
-      - .env
     environment:
-      CELERY_CMD: celery-cmd
-      IS_CELERY: 'True'
+      - IS_CELERY=True
+    entrypoint: ["/usr/src/{{project_name}}/entrypoint.sh"]
+    command: "celery-cmd"
 
+  # Nginx is serving django static and media files and proxies to django and geonode
   geonode:
     image: geonode/nginx:3.x
     build: ./docker/nginx/
-    restart: unless-stopped
     container_name: nginx4${COMPOSE_PROJECT_NAME}
-    stdin_open: true
-    # tty: true
-    labels:
-        org.geonode.component: nginx
-        org.geonode.instance.name: geonode
-    # depends_on:
-    #   - django
-    #   - geoserver
+    environment:
+      - HTTPS_HOST=${HTTPS_HOST}
+      - HTTP_HOST=${HTTP_HOST}
+      - HTTPS_PORT=${HTTPS_PORT}
+      - HTTP_PORT=${HTTP_PORT}
+      - LETSENCRYPT_MODE=${LETSENCRYPT_MODE}
+      - RESOLVER=127.0.0.11
     ports:
-      - "80:80"
-      - "443:443"
+      - "${HTTP_PORT}:80"
+      - "${HTTPS_PORT}:443"
     volumes:
       - nginx-confd:/etc/nginx
       - nginx-certificates:/geonode-certificates
       - statics:/mnt/volumes/statics
-    env_file:
-      - .env
+    restart: on-failure
 
   # Gets and installs letsencrypt certificates
   letsencrypt:
     image: geonode/letsencrypt:latest
     build: ./docker/letsencrypt/
-    restart: on-failure
+    container_name: letsencrypt4${COMPOSE_PROJECT_NAME}
+    environment:
+      - HTTPS_HOST=${HTTPS_HOST}
+      - HTTP_HOST=${HTTP_HOST}
+      - ADMIN_EMAIL=${ADMIN_EMAIL}
+      - LETSENCRYPT_MODE=${LETSENCRYPT_MODE}
     volumes:
-      - nginx-certificates:/geonode-certificates/
+      - nginx-certificates:/geonode-certificates
+    restart: on-failure
+
+  # Geoserver backend
+  geoserver:
+    image: geonode/geoserver:2.18.2
+    build: ./docker/geoserver/
+    container_name: geoserver4${COMPOSE_PROJECT_NAME}
+    healthcheck:
+      test: "curl --fail --silent --write-out 'HTTP CODE : %{http_code}\n' --output /dev/null http://127.0.0.1:8080/geoserver/rest/workspaces/geonode.html"
+      interval: 60s
+      timeout: 10s
+      retries: 1
+      start_period: 60s
     env_file:
       - .env
+    volumes:
+      - statics:/mnt/volumes/statics
+      - geoserver-data-dir:/geoserver_data/data
+      - backup-restore:/backup_restore
+      - data:/data
+      - tmp:/tmp
+    restart: on-failure
 
   data-dir-conf:
     image: geonode/geoserver_data:2.18.2
-    restart: on-failure
     container_name: gsconf4${COMPOSE_PROJECT_NAME}
-    labels:
-        org.geonode.component: conf
-        org.geonode.instance.name: geonode
     command: /bin/true
     volumes:
       - geoserver-data-dir:/geoserver_data/data
+    restart: on-failure
+
+  # PostGIS database.
+  db:
+    # use geonode official postgis 13 image
+    image: geonode/postgis:13
+    command: postgres -c "max_connections=${POSTGRESQL_MAX_CONNECTIONS}"
+    container_name: db4${COMPOSE_PROJECT_NAME}
+    env_file:
+      - .env
+    volumes:
+      - dbdata:/var/lib/postgresql/data
+      - dbbackups:/pg_backups
+    restart: on-failure
+    # uncomment to enable remote connections to postgres
+    #ports:
+    #  - "5432:5432"
+
+  # Vanilla RabbitMQ service. This is needed by celery
+  rabbitmq:
+    image: rabbitmq:3.7-alpine
+    container_name: rabbitmq4${COMPOSE_PROJECT_NAME}
+    volumes:
+      - rabbitmq:/var/lib/rabbitmq
+    restart: on-failure
 
   jenkins:
     image: jenkins/jenkins:lts
@@ -182,3 +174,5 @@ volumes:
     name: ${COMPOSE_PROJECT_NAME}-data
   tmp:
     name: ${COMPOSE_PROJECT_NAME}-tmp
+  rabbitmq:
+    name: ${COMPOSE_PROJECT_NAME}-rabbitmq
