--- project_name/templates/geonode_base.html
@@ -1,4 +1,0 @@
-{% extends "site_base.html" %}
-
-{% load i18n %}
-{% load staticfiles %}
