+++ src/README.md
@@ -1,0 +1,1 @@
+# {{ project_name|title }}
