+++ src/fixtures/default_oauth_apps.json
