+++ src/fixtures/default_oauth_apps_docker.json
