+++ src/fixtures/django_celery_beat.json
