+++ src/pavement.py
@@ -612,11 +612,11 @@ def start_django(options):
     """
     settings = options.get('settings', '')
     if settings and 'DJANGO_SETTINGS_MODULE' not in settings:
-        settings = 'DJANGO_SETTINGS_MODULE=%s' % settings
+        settings = f'DJANGO_SETTINGS_MODULE={settings}'
     bind = options.get('bind', '0.0.0.0:8000')
     port = bind.split(":")[1]
     foreground = '' if options.get('foreground', False) else '&'
-    sh('%s python -W ignore manage.py runserver %s %s' % (settings, bind, foreground))
+    sh(f'{settings} python -W ignore manage.py runserver {bind} {foreground}')
 
     if ASYNC_SIGNALS:
         scheduler = '--statedb=worker.state -s celerybeat-schedule'
@@ -628,7 +628,7 @@ def start_django(options):
         sh(f'{settings} python -W ignore manage.py runmessaging {foreground}')
 
     # wait for Django to start
-    started = waitfor("http://localhost:" + port)
+    started = waitfor(f"http://localhost:{port}")
     if not started:
         info('Django never started properly or timed out.')
         sys.exit(1)
@@ -1129,7 +1129,7 @@ def publish(options):
         sh('git push origin %s' % version)
         sh('git tag -f debian/%s' % simple_version)
         sh('git push origin debian/%s' % simple_version)
-        # sh('git push origin 3.2.x')
+        # sh('git push origin 3.3.x')
         sh('python setup.py sdist upload -r pypi')
