+++ src/project_name/static/package.json
