+++ src/project_name/templates/geonode_base.html
@@ -1,0 +1,4 @@
+{% extends "site_base.html" %}
+
+{% load i18n %}
+{% load staticfiles %}
