+++ src/project_name/wsgi.py
