+++ src/scripts/misc/cleanup_pyc.sh
